Implement cache for stack templates
Cache is used to track which stacks were created from which stack
template in order to speed up event conditions that depend on them.
Cache is intended to work only on server side, as its purpose is to
speed up the event system.
Cache is cleared and set during scenario loading. Cache is updated each
time stack is removed or source template id is assigned.