Implement resource market site
Resource market is a new site object for stacks to visit
on a strategic map. Resource market allows player to exchange
gold and mana resources according to exchange rates. Rates
are defined using Lua scripts and can be specific for an object or
use default rules that are global for the whole game.